Create README.md
Write a function def uppertrianglegrayscale( im ): which receives an image file and converts only the upper triangle half of the image to gray scale. To access pixels p[i,j] is in the upper triangle the inner for loop over the counter j must run from 0 to int(m*i) with the diagonal slope m = (yres-1)/( xres-1) computed before the double loop.